removed discord, changed twitter to grincouncil
ideally @grincouncil twitter would follow @grinmw.
maybe @grinmw deserves official page listing since it is a narrow  band and we have a couple council members moderating, but I can't in clear conscience let the discord be listed as official (I have already seen it referenced as such) when it is such an attack vector, and then at the same time try to pass my twitter as official. would be an honor, but if that's what it takes to remove the discord official link then it must be done.