change twitter to grincouncil, remove discord link
I have already seen the discord referenced as official in other channels (and a random opinion touted as official), and I think it is too wide an attack surface and liability to have it listed on the main site. it is a great community, but lots of trash talking, etc. would be too easy for it to be comandeered and used to pitch scams or alternative projects. the intentions of the speculators who founded and moderate the channel are not as pure as the core grin team, and the link on the main grin site does not reflect this fact.

I would love for my twitter @grinmw to be listed as official, but can't in clear conscience have it, as technically an unofficial channel, listed on the main site but deny the right of the discord. the twitter is narrow band and has a couple of core members with full moderating permissions (@lehnberg and @yeastplume), so it might be seen as less of a liabilty, but still gives undue power to someone not on the council (me.. which could  be seen as a plus but what if I lose my mind or die). would love if the "official" twitter followed the grinmw account though. and if others think it's deserving so be it, but the discord cannot be listed on the main page in my opinion. it is already listed elsewhere on the wiki I believe, which is appropriate and productive.