Update Version on main index page
It seems like the release version is a unnecessary detail to keep in sync and updated so prominently on the main landing page.  Having the version incorrect (as it currently is) causes significant legitimacy issues for the new grin.mw website.  As a non-controversial temporary fix... this pull request fixes the version number.  As a long term solutions, I recommend removing this detail from the landing page.